+++ backend/contracts/utils/KojoStorage.sol
@@ -4,28 +4,31 @@ pragma solidity ^0.8.9;
 import "../access/KojoOwnable.sol";
 
 contract KojoStorage is KojoOwnable {
-    uint private startCapital;
-    uint private tokenSensitivity;
-    uint private wateringCost;
+    uint256 private startCapital;
+    uint256 private tokenSensitivity;
+    uint256 private wateringCost;
 
-    constructor(){
+    constructor() {
         startCapital = 100;
         tokenSensitivity = 1;
         wateringCost = 1;
     }
 
     // Allows owner to update the start capital given to users.
-    function handleUpdateStartCapital(uint _startCapital) public isOwner {
+    function handleUpdateStartCapital(uint256 _startCapital) public isOwner {
         startCapital = _startCapital;
     }
 
     // Allows owner to update how many tokens are distributed for a given percentage point.
-    function handleUpdateTokenSensitivity(uint _tokenSensitivity) public isOwner {
+    function handleUpdateTokenSensitivity(uint256 _tokenSensitivity)
+        public
+        isOwner
+    {
         tokenSensitivity = _tokenSensitivity;
     }
 
     // Allows owner to update the cost of watering a seed/plant.
-    function handleUpdateWateringCost(uint _wateringCost) public isOwner {
+    function handleUpdateWateringCost(uint256 _wateringCost) public isOwner {
         wateringCost = _wateringCost;
     }
 }
